Refactored the jdbc connection sources to pull out base.
