+++ src/main/java/com/j256/ormlite/jdbc/JdbcConnectionSource.java
@@ -1,19 +1,13 @@
 package com.j256.ormlite.jdbc;
 
-import java.io.IOException;
 import java.sql.DriverManager;
 import java.sql.SQLException;
 import java.util.Properties;
 
 import com.j256.ormlite.db.DatabaseType;
-import com.j256.ormlite.jdbc.db.DatabaseTypeUtils;
 import com.j256.ormlite.logger.Logger;
-import com.j256.ormlite.logger.LoggerFactory;
-import com.j256.ormlite.misc.IOUtils;
-import com.j256.ormlite.support.BaseConnectionSource;
 import com.j256.ormlite.support.ConnectionSource;
 import com.j256.ormlite.support.DatabaseConnection;
-import com.j256.ormlite.support.DatabaseConnectionProxyFactory;
 
 /**
  * Implementation of the ConnectionSource interface that supports what is needed by ORMLite. This is not thread-safe nor
@@ -27,17 +21,10 @@ import com.j256.ormlite.support.DatabaseConnectionProxyFactory;
  * 
  * @author graywatson
  */
-public class JdbcConnectionSource extends BaseConnectionSource implements ConnectionSource {
+public class JdbcConnectionSource extends BaseJdbcConnectionSource implements ConnectionSource {
 
-	private static Logger logger = LoggerFactory.getLogger(JdbcConnectionSource.class);
-
-	private String url;
 	private String username;
 	private String password;
-	protected DatabaseConnection connection;
-	protected DatabaseType databaseType;
-	protected boolean initialized = false;
-	private static DatabaseConnectionProxyFactory connectionProxyFactory;
 
 	/**
 	 * Constructor for Spring type wiring if you are using the set methods. If you are using Spring then your should
@@ -118,122 +105,12 @@ public class JdbcConnectionSource extends BaseConnectionSource implements Connec
 	 */
 	protected JdbcConnectionSource(String url, String username, String password, DatabaseType databaseType,
 			boolean initialize) throws SQLException {
-		this.url = url;
+		super(url, databaseType);
 		this.username = username;
 		this.password = password;
-		this.databaseType = databaseType;
 		if (initialize) {
 			initialize();
 		}
-	}
-
-	/**
-	 * Initialize the class after the setters have been called. If you are using the no-arg constructor and Spring type
-	 * wiring, this should be called after all of the set methods.
-	 * 
-	 * @throws SQLException
-	 *             If the driver associated with the database URL is not found in the classpath.
-	 */
-	public void initialize() throws SQLException {
-		if (initialized) {
-			return;
-		}
-		if (url == null) {
-			throw new SQLException("url was never set on " + getClass().getSimpleName());
-		}
-		if (databaseType == null) {
-			databaseType = DatabaseTypeUtils.createDatabaseType(url);
-		}
-		databaseType.loadDriver();
-		databaseType.setDriver(DriverManager.getDriver(url));
-		initialized = true;
-	}
-
-	@Override
-	public void close() throws IOException {
-		if (!initialized) {
-			throw new IOException(getClass().getSimpleName() + " was not initialized properly");
-		}
-		if (connection != null) {
-			connection.close();
-			logger.debug("closed connection #{}", connection.hashCode());
-			connection = null;
-		}
-	}
-
-	@Override
-	public void closeQuietly() {
-		IOUtils.closeQuietly(this);
-	}
-
-	public String getUrl() {
-		return url;
-	}
-
-	public void setUrl(String url) {
-		this.url = url;
-	}
-
-	@Override
-	public DatabaseConnection getReadOnlyConnection(String tableName) throws SQLException {
-		if (!initialized) {
-			throw new SQLException(getClass().getSimpleName() + " was not initialized properly");
-		}
-		return getReadWriteConnection(tableName);
-	}
-
-	@Override
-	public DatabaseConnection getReadWriteConnection(String tableName) throws SQLException {
-		if (!initialized) {
-			throw new SQLException(getClass().getSimpleName() + " was not initialized properly");
-		}
-		if (connection != null) {
-			if (connection.isClosed()) {
-				throw new SQLException("Connection has already been closed");
-			} else {
-				return connection;
-			}
-		}
-		connection = makeConnection(logger);
-		return connection;
-	}
-
-	@Override
-	public void releaseConnection(DatabaseConnection connection) throws SQLException {
-		if (!initialized) {
-			throw new SQLException(getClass().getSimpleName() + " was not initialized properly");
-		}
-		// noop right now
-	}
-
-	@Override
-	@SuppressWarnings("unused")
-	public boolean saveSpecialConnection(DatabaseConnection connection) throws SQLException {
-		// noop since this is a single connection source
-		return true;
-	}
-
-	@Override
-	public void clearSpecialConnection(DatabaseConnection connection) {
-		// noop since this is a single connection source
-	}
-
-	@Override
-	public DatabaseType getDatabaseType() {
-		if (!initialized) {
-			throw new IllegalStateException(getClass().getSimpleName() + " was not initialized properly");
-		}
-		return databaseType;
-	}
-
-	@Override
-	public boolean isOpen(String tableName) {
-		return connection != null;
-	}
-
-	@Override
-	public boolean isSingleConnection(String tableName) {
-		return true;
 	}
 
 	// not required
@@ -246,24 +123,7 @@ public class JdbcConnectionSource extends BaseConnectionSource implements Connec
 		this.password = password;
 	}
 
-	// not required
-	public void setDatabaseType(DatabaseType databaseType) {
-		this.databaseType = databaseType;
-	}
-
-	/**
-	 * Set to enable connection proxying. Set to null to disable.
-	 */
-	public static void setDatabaseConnectionProxyFactory(DatabaseConnectionProxyFactory connectionProxyFactory) {
-		JdbcConnectionSource.connectionProxyFactory = connectionProxyFactory;
-	}
-
-	/**
-	 * Make a connection to the database.
-	 * 
-	 * @param logger
-	 *            This is here so we can use the right logger associated with the sub-class.
-	 */
+	@Override
 	protected DatabaseConnection makeConnection(Logger logger) throws SQLException {
 		Properties properties = new Properties();
 		if (username != null) {
@@ -275,9 +135,6 @@ public class JdbcConnectionSource extends BaseConnectionSource implements Connec
 		DatabaseConnection connection = new JdbcDatabaseConnection(DriverManager.getConnection(url, properties));
 		// by default auto-commit is set to true
 		connection.setAutoCommit(true);
-		if (connectionProxyFactory != null) {
-			connection = connectionProxyFactory.createProxy(connection);
-		}
 		logger.debug("opened connection to {} got #{}", url, connection.hashCode());
 		return connection;
 	}
